requirements.txt
requests
streamlit
langchain
python-docx
streamlit-lottie
